Create README.md
I have a specialty in setting up the Windows operating system very quickly. I dedicate myself to reading tricks to speed up the system and which programs are the best on the market. I also learned to combine the parts of the computer for better performance. I have experience since 2004, and so far all the clients I have had have been very satisfied with my work.